requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
abseil-cpp=20211102.0=hd4dd3e8_0
absl-py=1.3.0=py38h06a4308_0
aiohttp=3.8.3=py38h5eee18b_0
aiosignal=1.2.0=pyhd3eb1b0_0
ale-py=0.7.4=pypi_0
appdirs=1.4.4=pyhd3eb1b0_0
asciinema=2.2.0=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=4.0.2=py38h06a4308_0
attrs=22.1.0=py38h06a4308_0
autopep8=1.6.0=pyhd3eb1b0_1
autorom=0.6.1=pypi_0
autorom-accept-rom-license=0.6.1=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
blinker=1.4=py38h06a4308_0
bottleneck=1.3.5=py38h7deecbd_0
box2d=2.3.10=pypi_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.0=h5eee18b_0
ca-certificates=2023.05.30=h06a4308_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2023.5.7=py38h06a4308_0
cffi=1.15.1=py38h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py38h06a4308_0
cloudpickle=2.2.1=pyhd8ed1ab_0
contourpy=1.0.5=py38hdb19cb5_0
cryptography=39.0.1=py38h9ce1e76_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py38h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
entrypoints=0.4=py38h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.4.9=h6a678d5_0
farama-notifications=0.0.4=pyhd8ed1ab_0
ffmpeg=4.2.2=h20bf706_0
filelock=3.9.0=py38h06a4308_0
fontconfig=2.14.1=h4c34cd2_2
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=py38h5eee18b_0
giflib=5.2.1=h5eee18b_3
glib=2.69.1=he621ea3_2
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py38heeb90bb_0
gnutls=3.6.15=he1e5248_0
google-auth=2.6.0=pyhd3eb1b0_0
google-auth-oauthlib=0.5.2=py38h06a4308_0
grpc-cpp=1.48.2=h5bf31a4_0
grpcio=1.48.2=py38h5bf31a4_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
gym=0.21.0=pypi_0
gym-aima=0.0.1=pypi_0
gym-bandits=0.1=pypi_0
gym-walk=0.0.2=pypi_0
icu=58.2=he6710b0_3
idna=3.4=py38h06a4308_0
ignite=0.4.12=py_0
importlib-metadata=4.13.0=pypi_0
importlib_metadata=1.5.0=py38_0
importlib_resources=5.2.0=pyhd3eb1b0_1
intel-openmp=2023.1.0=hdb19cb5_46305
ipykernel=6.15.0=pyh210e3f2_0
ipython=8.12.0=py38h06a4308_0
jax-jumpy=1.0.0=pyhd8ed1ab_0
jedi=0.18.1=py38h06a4308_1
jinja2=3.1.2=py38h06a4308_0
joblib=1.2.0=pypi_0
jpeg=9e=h5eee18b_1
jupyter_client=7.0.6=pyhd8ed1ab_0
jupyter_core=5.3.0=py38h06a4308_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.4=py38h6a678d5_0
krb5=1.19.4=h568e23c_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.6.1.9=0
libcurand=10.3.2.106=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_0
libedit=3.1.20221030=h5eee18b_0
libevent=2.1.12=h8f2d780_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libllvm10=10.0.1=he513fc3_3
libllvm14=14.0.6=hdb19cb5_3
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libopus=1.3.1=h7b6447c_0
libpng=1.6.39=h5eee18b_0
libpq=12.9=h16c4e8d_3
libprotobuf=3.20.3=he621ea3_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libvpx=1.7.0=h439df22_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h5eee18b_1
libxml2=2.10.3=hcbfbd50_0
libxslt=1.1.37=h2085143_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.3=pyhd8ed1ab_0
markdown-it-py=2.2.0=pypi_0
markupsafe=2.1.1=py38h7f8727e_0
matplotlib=3.7.1=py38h578d9bd_0
matplotlib-base=3.7.1=py38h417a72b_1
matplotlib-inline=0.1.6=py38h06a4308_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h6d00ec8_46342
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.6=py38h417a72b_1
mkl_random=1.2.2=py38h417a72b_1
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.2.1=py38h06a4308_0
multidict=6.0.2=py38h5eee18b_0
munkres=1.1.4=py_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=py38h06a4308_0
nettle=3.7.3=hbbd107a_1
networkx=2.8.4=py38h06a4308_1
nspr=4.35=h6a678d5_0
nss=3.89.1=h6a678d5_0
numexpr=2.8.4=py38hc78ab66_1
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
oauthlib=3.2.2=py38h06a4308_0
opencv-python=4.1.2.30=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1u=h7f8727e_0
packaging=23.0=py38h06a4308_0
pandas=1.5.2=py38h417a72b_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.4.0=py38h6a678d5_0
pip=23.1.2=pypi_0
platformdirs=2.5.2=py38h06a4308_0
ply=3.11=py38_0
pooch=1.4.0=pyhd3eb1b0_0
prompt-toolkit=3.0.36=py38h06a4308_0
prompt_toolkit=3.0.36=hd3eb1b0_0
protobuf=3.20.3=py38h6a678d5_0
psutil=5.9.0=py38h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pybullet=3.2.5=pypi_0
pycodestyle=2.10.0=py38h06a4308_0
pycparser=2.21=pyhd3eb1b0_0
pyglet=1.5.27=pypi_0
pyglm=2.7.0=pypi_0
pygments=2.15.1=py38h06a4308_1
pyjwt=2.4.0=py38h06a4308_0
pyopenssl=23.0.0=py38h06a4308_0
pyparsing=3.0.9=py38h06a4308_0
pyqt=5.15.7=py38h6a678d5_1
pyqt5-sip=12.11.0=py38h6a678d5_1
pyside2=5.15.2.1=pypi_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.16=h7a1cb2a_3
python-dateutil=2.8.2=pyhd3eb1b0_0
python-dotenv=1.0.0=pypi_0
python_abi=3.8=2_cp38
pytorch=2.0.1=py3.8_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pytz=2022.7=py38h06a4308_0
pyu2f=0.1.5=pyhd8ed1ab_0
pyzmq=25.1.0=py38h6a678d5_0
qt-main=5.15.2=h8373d8f_8
qt-webengine=5.15.9=hbbf29b9_6
qtwebkit=5.212=h3fafdc1_5
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
requests=2.29.0=py38h06a4308_0
requests-oauthlib=1.3.0=py_0
rich=13.4.1=pypi_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-learn=1.2.2=pypi_0
scipy=1.10.1=py38hf6e8229_1
setuptools=67.8.0=py38h06a4308_0
shiboken2=5.15.2.1=pypi_0
sip=6.6.2=py38h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h5eee18b_0
stable-baselines3=1.8.0=pypi_0
stack_data=0.2.0=pyhd3eb1b0_0
swig=4.1.1=pypi_0
sympy=1.11.1=py38h06a4308_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.13.0=pyhd8ed1ab_0
tensorboard-data-server=0.7.0=py38h52d8a92_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
torchaudio=2.0.2=py38_cu118
torchtriton=2.0.0=py38
torchvision=0.15.2=py38_cu118
tornado=5.1.1=pypi_0
tqdm=4.65.0=pypi_0
traitlets=5.7.1=py38h06a4308_0
typing_extensions=4.6.3=py38h06a4308_0
urllib3=1.26.16=py38h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
werkzeug=2.2.3=py38h06a4308_0
wheel=0.38.4=py38h06a4308_0
x264=1!157.20191217=h7b6447c_0
xz=5.4.2=h5eee18b_0
yarl=1.8.1=py38h5eee18b_0
zeromq=4.3.4=h2531618_0
zipp=3.11.0=py38h06a4308_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
